requirements.txt
Flask==2.3.3
python-telegram-bot[webhooks]==20.3
httpx==0.24.1
openai==1.30.0
